Refactor/test components cmake (#8677)
* working

* working

* working

* refactor

* minor improvements

* windows VS 64 bits
